Moving pt_BR CNPJ generator to Company provider
- CNPJ is a company document, thus the move
- check digit function was properly named, and moved into a separate
  file, as it's used by both classes. Furthermore, as the check digit
  algorithm is particular to Brazil (it's a variation of mod11, with
  different maximum multipliers given the number length) it was not
  placed in Faker/Calculator.
