+++ src/Faker/Provider/pt_BR/Company.php
@@ -2,6 +2,8 @@
 
 namespace Faker\Provider\pt_BR;
 
+require_once "check_digit.php";
+
 class Company extends \Faker\Provider\Company
 {
     protected static $formats = array(
@@ -13,4 +15,20 @@ class Company extends \Faker\Provider\Company
     );
 
     protected static $companySuffix = array('e Filho', 'e Filha', 'e Filhos', 'e Associados', 'e Flia.', 'SRL', 'SA', 'S. de H.');
+
+    /**
+     * A random CNPJ number.
+     * @link http://en.wikipedia.org/wiki/CNPJ
+     * @link http://pt.wikipedia.org/wiki/CNPJ#Algoritmo_de_Valida.C3.A7.C3.A3o
+     * @param bool $formatted If the number should have dots/slashes/dashes or not.
+     * @return string
+     */
+    public function cnpj($formatted = true)
+    {
+        $n = $this->generator->numerify('########0001');
+        $n .= check_digit($n);
+        $n .= check_digit($n);
+
+        return $formatted? vsprintf('%d%d.%d%d%d.%d%d%d/%d%d%d%d-%d%d', str_split($n)) : $n;
+    }
 }
